Revisit kernels (#55)
* first commit

* address compilation issues

* change signatures

* add changes

* add changes

* add changes

* refactor cache manager copy blocks kernels

* futher refactor of cache_manager logic

* refactor cache manager tests

* refactor copy blocks kernel

* small refactor for copy blocks ffi

* refactor cache manager copy blocks kernels

* refactor cache manager copy blocks kernels

* refactor changes

* additional device checks

* also check for sm device

* add stream synchronization for copy_blocks and reshape_and_cache_flash methods

+++ csrc/kernels/flash.h
@@ -16,7 +16,7 @@ constexpr int D_DIM = 2;
 ////////////////////////////////////////////////////////////////////////////////////////////////////
 
 struct Qkv_params {
-    using index_t = int64_t;
+    using index_t = uint32_t;
 
     // The QKV matrices.
     void *__restrict__ q_ptr;
